requirements.txt
python-telegram-bot>=20.0
google-genai
python-dotenv
Flask
